+++ service/controller/managementcluster/resource.go
@@ -11,8 +11,7 @@ import (
 	vpa_clientset "k8s.io/autoscaler/vertical-pod-autoscaler/pkg/client/clientset/versioned"
 
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/alertmanager"
-	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/alertmanagerconfigsecret"
-	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/alertmanagerwiring"
+	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/alertmanagerconfig"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/heartbeat"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/heartbeatrouting"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/heartbeatwebhookconfig"
@@ -138,14 +137,12 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 		}
 	}
 
-	var alertmanagerConfigSecretResource resource.Interface
+	var alertmanagerConfigResource resource.Interface
 	{
-		c := alertmanagerconfigsecret.Config{
-			K8sClient: config.K8sClient,
-			Logger:    config.Logger,
-
+		c := alertmanagerconfig.Config{
+			K8sClient:        config.K8sClient,
+			Logger:           config.Logger,
 			Installation:     config.Installation,
-			Pipeline:         config.Pipeline,
 			Provider:         config.Provider,
 			HTTPProxy:        config.HTTPProxy,
 			HTTPSProxy:       config.HTTPSProxy,
@@ -154,9 +151,10 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 			GrafanaAddress:   config.GrafanaAddress,
 			SlackApiURL:      config.SlackApiURL,
 			SlackProjectName: config.SlackProjectName,
+			Pipeline:         config.Pipeline,
 		}
 
-		alertmanagerConfigSecretResource, err = alertmanagerconfigsecret.New(c)
+		alertmanagerConfigResource, err = alertmanagerconfig.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -175,19 +173,6 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 		}
 
 		heartbeatWebhookConfigResource, err = heartbeatwebhookconfig.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	var alertmanagerWiringResource resource.Interface
-	{
-		c := alertmanagerwiring.Config{
-			K8sClient: config.K8sClient,
-			Logger:    config.Logger,
-		}
-
-		alertmanagerWiringResource, err = alertmanagerwiring.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -336,9 +321,8 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 		etcdCertificatesResource,
 		rbacResource,
 		alertmanagerResource,
-		alertmanagerConfigSecretResource,
+		alertmanagerConfigResource,
 		heartbeatWebhookConfigResource,
-		alertmanagerWiringResource,
 		scrapeConfigResource,
 		remoteWriteConfigResource,
 		prometheusResource,