+++ service/controller/resource/alerting/alertmanagerconfig/resource.go
@@ -1,9 +1,10 @@
-package alertmanagerconfigsecret
+package alertmanagerconfig
 
 import (
 	"io/ioutil"
 	"path"
 	"reflect"
+	"strings"
 
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
@@ -17,7 +18,7 @@ import (
 )
 
 const (
-	Name                     = "alertmanagerconfigsecret"
+	Name                     = "alertmanagerconfig"
 	templateDirectory        = "/opt/prometheus-meta-operator"
 	templatePath             = "files/templates/alertmanager/alertmanager.yaml"
 	notificationTemplatePath = "files/templates/alertmanager/notification-template.tmpl"
@@ -85,7 +86,7 @@ func New(config Config) (*generic.Resource, error) {
 
 func getObjectMeta(v interface{}, config Config) (metav1.ObjectMeta, error) {
 	return metav1.ObjectMeta{
-		Name:      "alertmanager-config",
+		Name:      key.AlertManagerSecretName(),
 		Namespace: key.NamespaceMonitoring(),
 	}, nil
 }
@@ -101,7 +102,7 @@ func toSecret(v interface{}, config Config) (*corev1.Secret, error) {
 		return nil, microerror.Mask(err)
 	}
 
-	alertmanagerConfig, err := toData(v, config)
+	alertmanagerConfigSecret, err := toData(v, config)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
@@ -109,9 +110,9 @@ func toSecret(v interface{}, config Config) (*corev1.Secret, error) {
 	secret := &corev1.Secret{
 		ObjectMeta: objectMeta,
 		Data: map[string][]byte{
-			key.AlertmanagerKey():         alertmanagerConfig,
-			"notification-templates.tmpl": notificationTemplate,
-			key.OpsgenieKey():             []byte(config.OpsgenieKey),
+			key.AlertmanagerKey():        alertmanagerConfigSecret,
+			"notification-template.tmpl": notificationTemplate,
+			key.OpsgenieKey():            []byte(config.OpsgenieKey),
 		},
 		Type: "Opaque",
 	}
@@ -140,12 +141,13 @@ func toData(v interface{}, config Config) ([]byte, error) {
 
 func getTemplateData(cluster metav1.Object, config Config) (*TemplateData, error) {
 	var proxyURL *string = nil
-	if len(config.HTTPSProxy) > 0 {
-		proxyURL = &config.HTTPSProxy
-	} else if len(config.HTTPProxy) > 0 {
-		proxyURL = &config.HTTPProxy
+	if !strings.Contains(config.NoProxy, "api.opsgenie.com") {
+		if len(config.HTTPSProxy) > 0 {
+			proxyURL = &config.HTTPSProxy
+		} else if len(config.HTTPProxy) > 0 {
+			proxyURL = &config.HTTPProxy
+		}
 	}
-
 	d := &TemplateData{
 		Provider:         config.Provider,
 		Installation:     config.Installation,